event (#5)
* event api fixed to check user role

* talk, speaker, tag and partner api fixed to check user role

* apis refactoring for every class

* talkRepository fix

* tag class fix

* repository fix

* id set with randomUUID

+++ src/main/java/fullstack/rest/resources/EventResource.java
@@ -1,5 +1,8 @@
 package fullstack.rest.resources;
 
+import fullstack.persistence.model.Role;
+import fullstack.service.UserService;
+import fullstack.service.exception.UserNotFoundException;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import fullstack.persistence.model.Event;
@@ -8,6 +11,7 @@ import fullstack.persistence.model.Talk;
 import fullstack.service.EventService;
 import fullstack.service.SpeakerService;
 import fullstack.service.TalkService;
+import jakarta.ws.rs.core.Response;
 
 import java.util.List;
 
@@ -17,11 +21,13 @@ public class EventResource {
     private final EventService eventService;
     private final TalkService talkService;
     private final SpeakerService speakerService;
+    private final UserService userService;
 
-    public EventResource(EventService eventService, TalkService talkService, SpeakerService speakerService) {
+    public EventResource(EventService eventService, TalkService talkService, SpeakerService speakerService, UserService userService) {
         this.eventService = eventService;
         this.talkService = talkService;
         this.speakerService = speakerService;
+        this.userService = userService;
     }
 
     @GET
@@ -61,21 +67,40 @@ public class EventResource {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Event createEvent(Event event) {
-        return eventService.save(event);
+    public Response createEvent(@CookieParam("sessionId") String sessionId, Event event) throws UserNotFoundException {
+        Role userRole = userService.getUserRoleBySessionId(sessionId);
+        if (userRole != Role.admin) {
+            return Response.status(Response.Status.FORBIDDEN).entity("Access denied").build();
+        }
+        Event savedEvent = eventService.save(event);
+        return Response.ok(savedEvent).build();
     }
 
     @DELETE
     @Path("/{id}")
-    public void deleteEvent(@PathParam("id") String id) {
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response deleteEvent(@CookieParam("sessionId") String sessionId, @PathParam("id") String id) throws UserNotFoundException {
+        Role userRole = userService.getUserRoleBySessionId(sessionId);
+        if (userRole != Role.admin) {
+            return Response.status(Response.Status.FORBIDDEN).entity("Access denied").build();
+        }
         eventService.deleteById(id);
+        return Response.noContent().build();
     }
 
     @PUT
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public int updateEvent(@PathParam("id") String id, Event event) {
-        return eventService.update(id, event);
+    public Response updateEvent(@CookieParam("sessionId") String sessionId, @PathParam("id") String id, Event event) throws UserNotFoundException {
+        Role userRole = userService.getUserRoleBySessionId(sessionId);
+        if (userRole != Role.admin) {
+            return Response.status(Response.Status.FORBIDDEN).entity("Access denied").build();
+        }
+        int updated = eventService.update(id, event);
+        if (updated == 0) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Event not found").build();
+        }
+        return Response.ok().build();
     }
 }
